OpenSubtitles: Moving to own namespace and folder
