+++ Extensions/PlayerExtensions/OpenSubtitles/OpenSubtitles.LanguageHandler.cs
@@ -14,12 +14,13 @@
 // You should have received a copy of the GNU Lesser General Public
 // License along with this library.
 // 
+
 using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Windows.Forms;
 
-namespace Mpdn.Extensions.PlayerExtensions
+namespace Mpdn.Extensions.PlayerExtensions.OpenSubtitles
 {
     public class OpenSubtitlesLanguageHandler
     {
