+++ Extensions/PlayerExtensions/OpenSubtitles/OpenSubtitles.PlayerExtension.cs
@@ -14,16 +14,17 @@
 // You should have received a copy of the GNU Lesser General Public
 // License along with this library.
 // 
+
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
+using System.IO;
 using Mpdn.Extensions.Framework;
 using Mpdn.Extensions.Framework.Controls;
-using System.IO;
 using Mpdn.Extensions.PlayerExtensions.Exceptions;
 
-namespace Mpdn.Extensions.PlayerExtensions
+namespace Mpdn.Extensions.PlayerExtensions.OpenSubtitles
 {
     public class OpenSubtitlesExtension : PlayerExtension<OpenSubtitlesSettings, OpenSubtitlesConfigDialog>
     {
