+++ Mpdn.Extensions.csproj
@@ -133,18 +133,18 @@
     </Compile>
     <Compile Include="Extensions\PlayerExtensions\MouseControl.cs" />
     <Compile Include="Extensions\PlayerExtensions\Navigation.cs" />
-    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles.ConfigDialog.cs">
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\OpenSubtitles.ConfigDialog.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles.ConfigDialog.designer.cs">
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\OpenSubtitles.ConfigDialog.designer.cs">
       <DependentUpon>OpenSubtitles.ConfigDialog.cs</DependentUpon>
     </Compile>
-    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles.Downloader.cs" />
-    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles.PlayerExtension.cs" />
-    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles.Subtitles.Form.cs">
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\OpenSubtitles.Downloader.cs" />
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\OpenSubtitles.PlayerExtension.cs" />
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\OpenSubtitles.Subtitles.Form.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles.Subtitles.Form.Designer.cs">
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\OpenSubtitles.Subtitles.Form.Designer.cs">
       <DependentUpon>OpenSubtitles.Subtitles.Form.cs</DependentUpon>
     </Compile>
     <Compile Include="Extensions\PlayerExtensions\Playlist.cs" />
@@ -344,7 +344,7 @@
     </Compile>
     <Compile Include="Extensions\RenderScripts\Shiandow.SuperRes.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles.LanguageHandler.cs" />
+    <Compile Include="Extensions\PlayerExtensions\OpenSubtitles\OpenSubtitles.LanguageHandler.cs" />
     <None Include="Extensions\RenderScripts\ChromaScaler\Chroma.hlsl" />
     <None Include="Extensions\RenderScripts\Common\ColourProcessing.hlsl" />
     <None Include="Extensions\RenderScripts\Common\GammaToLab.hlsl" />
